ghci: split actions into separate workflow
Split the ghci-actions tests into a separate workflow. They are not
related to the container creation, so lets keep them separate.

This also allows us to run them on every PR and PUSH, rather than just
on selective commits. There is no harm in running them and they are
pretty fast, anyway.