ghci: integrate into repository
This integrates the GHCI containers into the `osbuild/containers`
repository and drops all the `GHCI` prefixes and indirections. The
`ghci` directory is now renamed to `src` and allows for all kinds of
containers and infrastructure, not just for GitHub-CI.

The CI logic is also modified to serve our needs better. The
image-selector now works for all targets, the `rc` logic now always
picks the `rc` name, and the `v*` logic now picks the specified name,
but always aliases with the commit-sha.

This way, we do not produce needless images, but instead only ever keep
the releases around. We can then easily find previous/older releases.